Combine DbContextOperations tests
